moved the boards and their cards and styles into each respective folders. shrinking the screen size gets me a bg green col or, but if i make it bigger it doesn't go back.
