+++ src/components/RoomView/boardAndCardsOperatives/cardStyles.ts
