+++ src/components/RoomView/boardAndCardsSpy/spyCardStyles.ts
@@ -155,5 +155,7 @@ const blackCardStyles = {
   back: blackRevealed,
 };
 
+// Do media query
+
 const allCardStyles = { redCardStyles, blueCardStyles, beigeCardStyles, blackCardStyles };
 export default allCardStyles;
